git-svn-id: https://omb-dashboardplatform.reisys.com:10039/NYC_Checkbook_2_0/Database@1492 948744eb-f611-4f28-8820-250fb6d8f47d
Former-commit-id: b157e4ea60272bf06ffa767bd8a467c5364d146e

+++ checkbook/Data Model/NYCCheckbookETL_DDL.sql
@@ -3102,7 +3102,7 @@ CREATE TABLE stg_fms_accounting_line(
 	location_history_id int,
 	agreement_id bigint,
 	masked_agency_history_id smallint,
-	masked_department_history_id smallint,
+	masked_department_history_id int,
 	uniq_id bigint default nextval('seq_stg_fms_accounting_line_uniq_id'),
 	invalid_flag char(1),
 	invalid_reason varchar		)
